Removing references to unused google_analytics properties
When you forked my repo, you also copied my google analytics properties. Please remove them so that it is not interfering with my google analytics statistics.

If your interest is to build a blog, I would recommend following the blog post as written, which provides a template you can use to create a blog. Forking my repo is not the correct way to do it.

https://chadbaldwin.net/2021/03/14/how-to-build-a-sql-blog.html